Test is now running and numerical field changed from int to long

+++ common/common/src/main/java/com/trifork/stamdata/persistence/RecordPersister.java
@@ -95,7 +95,7 @@ public class RecordPersister
                 }
                 else if (fieldSpecification.type == RecordSpecification.RecordFieldType.NUMERICAL)
                 {
-                    preparedStatement.setInt(index, (Integer) record.get(fieldSpecification.name));
+                    preparedStatement.setLong(index, (Long) record.get(fieldSpecification.name));
                 }
                 else
                 {
