+++ nsp/batch-copy-ws/src/main/java/dk/nsi/stamdata/replication/webservice/RecordXmlGenerator.java
@@ -134,7 +134,7 @@ public class RecordXmlGenerator
         }
         else if (fieldSpecification.type == RecordFieldType.NUMERICAL)
         {
-            return Integer.toString((Integer) record.get(fieldSpecification.name));
+            return Long.toString((Long) record.get(fieldSpecification.name));
         }
         else
         {
